services/analyst/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
openai
python-dotenv